--- src/app/edit-task/edit-task.component.html
@@ -1,10 +1,0 @@
-<!--<h3>id: {{id}}</h3>-->
-<!--<form class="flex flex-row items-center justify-evenly p-4">-->
-<!--  <mat-form-field>-->
-<!--    <mat-label>nowa nazwa twojego zadania</mat-label>-->
-<!--      <input matInput type="text" [formControl]="taskEditControl" #newDescription >-->
-
-<!--  </mat-form-field>-->
-
-<!--  <button mat-flat-button color="primary" class="ml-4" (click)="onEdit(newDescription.value)" >zatwierdz</button>-->
-<!--</form>-->
