--- src/app/edit-task/edit-task.component.spec.ts
@@ -1,21 +1,0 @@
-// import { ComponentFixture, TestBed } from '@angular/core/testing';
-// import { EditTaskComponent } from './edit-task.component';
-// describe('EditTaskComponent', () => {
-//   let component: EditTaskComponent;
-//   let fixture: ComponentFixture<EditTaskComponent>;
-//
-//   beforeEach(async () => {
-//     await TestBed.configureTestingModule({
-//       declarations: [ EditTaskComponent ]
-//     })
-//     .compileComponents();
-//
-//     fixture = TestBed.createComponent(EditTaskComponent);
-//     component = fixture.componentInstance;
-//     fixture.detectChanges();
-//   });
-//
-//   it('should create', () => {
-//     expect(component).toBeTruthy();
-//   });
-// });
