+++ src/app/login/login.component.ts
@@ -1,7 +1,7 @@
 import {Component, DoCheck, OnDestroy, OnInit} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
-import {UsersHandlerService} from "../users-handler.service";
-import {AuthService} from "../auth.service";
+import {UsersHandlerService} from "../services/users-handler.service";
+import {AuthService} from "../services/auth.service";
 import {map, Observable, Subject, switchMap, takeUntil, timer} from "rxjs";
 import {AbstractControl, AsyncValidatorFn, FormControl, ValidationErrors, Validators} from "@angular/forms";
 import {SignUpValidateService} from "../validators/sign-up-validate.service";
@@ -75,7 +75,6 @@ export class LoginComponent implements DoCheck, OnInit,OnDestroy{
 }
   );
 
-
   getErrorMessageSignUp() {
     if (this.signUpControl.hasError('required')) {
       return 'Nazwa użytkownika nie może być pusta!';
@@ -106,9 +105,6 @@ export class LoginComponent implements DoCheck, OnInit,OnDestroy{
     return '';
   }
 
-
-
-
   ngOnDestroy() {
     this.unSubAuthorize$.next();
     this.unSubAuthorize$.unsubscribe();
@@ -116,35 +112,5 @@ export class LoginComponent implements DoCheck, OnInit,OnDestroy{
     this.unSubAddNewUser$.next();
     this.unSubAddNewUser$.unsubscribe();
   }
-
-  // checkUsernameAvailability(http: HttpClient): AsyncValidatorFn {
-  //   return (control: AbstractControl): Observable<ValidationErrors | null> => {
-  //     const username = control.value;
-  //     return http.get<boolean>(`your_api_endpoint/check-username/${username}`).pipe(
-  //       map(isAvailable => (isAvailable ? null : { usernameTaken: true }))
-  //     );
-  //   };
-  // }
-  // checkUserArleadyExists(http: HttpClient): AsyncValidatorFn {
-  //   return (control: AbstractControl): Observable<ValidationErrors | null> => {
-  //     const username = control.value;
-  //     console.log("w checkusersalredyexists ");
-  //     return this.userHandler.authenticate(username).pipe(
-  //       map(exists => (exists ? { usernameTaken: true} : null ))
-  //     );
-  //   };
-  // }
-
-  // checkUserArleadyExists(http: HttpClient): AsyncValidatorFn {
-  //   return (control: AbstractControl): Observable<ValidationErrors | null> => {
-  //     const username = control.value;
-  //     const username2='krystiano'
-  //     console.log(`http://localhost:8080/login/${username}`);
-  //     return http.get<boolean>(`http://localhost:8080/login/${username2}`).pipe(
-  //       map(exists => (exists ? { usernameTaken: true } : null ))
-  //     );
-  //   };// 1. usatwic na sztywno nazwe, 2. usunac inne validatory
-  //   //nic nie działa. w takim razie sproboje zobaczy jak zroibli to fachowncy od angulara
-  // }
 
 }
