+++ src/app/validators/sign-up-validate.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import {AbstractControl, AsyncValidator, ValidationErrors} from "@angular/forms";
-import {UsersHandlerService} from "../users-handler.service";
+import {UsersHandlerService} from "../services/users-handler.service";
 import {Observable, of} from "rxjs";
 import {catchError, map} from "rxjs/operators";
 import {HttpClient} from "@angular/common/http";
